% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filter_too_many_missing}
\alias{filter_too_many_missing}
\title{Filter Genes with Too Many Missing Values}
\usage{
filter_too_many_missing(MS1.df, MS2.df, full.groups = 1)
}
\arguments{
\item{MS1.df}{A data frame containing MS1 quantification data.}

\item{MS2.df}{A data frame containing MS2 quantification data.}

\item{full.groups}{An integer specifying the minimum number of sample groups required to keep a gene.}
}
\value{
A vector of gene names that pass the filter.
}
\description{
This function filters out genes/proteins that have too many missing values in both MS1 and MS2 data frames.
}
\examples{
# Example usage:
# filtered_genes <- filter_too_many_missing(MS1.df, MS2.df, full.groups = 1)
}
