% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fit_DEqMS_model}
\alias{fit_DEqMS_model}
\title{Fit DEqMS Model}
\usage{
fit_DEqMS_model(data, contrast_list)
}
\arguments{
\item{data}{A SummarizedExperiment object containing quantification data.}

\item{contrast_list}{A list of contrasts to test.}
}
\value{
A DEqMS model fit object.
}
\description{
This function fits a limma & DEqMS model to the quantification data.
}
\examples{
# Example usage:
# fit <- fit_DEqMS_model(data, contrast_list)
}
