% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_MNAR}
\alias{get_MNAR}
\title{Get MNAR Genes}
\usage{
get_MNAR(df)
}
\arguments{
\item{df}{A data frame containing quantification data.}
}
\value{
A logical vector indicating whether each gene is MNAR.
}
\description{
This function identifies genes that are missing not at random (MNAR) in the data frame.
}
\examples{
# Example usage:
# MNAR <- get_MNAR(df)
}
