% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap}
\usage{
plot_heatmap(
  data.sig,
  color_scale,
  colors_columns,
  qvalues = NA,
  title = NA,
  max_rows = 20,
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  scale_by_row = TRUE
)
}
\arguments{
\item{data.sig}{A SummarizedExperiment object containing significant data.}

\item{color_scale}{A color scale for the heatmap.}

\item{colors_columns}{A list of colors for column annotations.}

\item{qvalues}{A vector of q-values for the data.}

\item{title}{A character string specifying the title of the heatmap.}

\item{max_rows}{An integer specifying the maximum number of rows to display.}

\item{cluster_rows}{A logical value indicating whether to cluster rows.}

\item{cluster_cols}{A logical value indicating whether to cluster columns.}

\item{scale_by_row}{A logical value indicating whether to scale data by row.}
}
\value{
A heatmap of significant data.
}
\description{
This function creates a heatmap of significant data.
}
\examples{
# Example usage:
# plot_heatmap(data.sig, color_scale, colors_columns)
}
