% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_imputation_density}
\alias{plot_imputation_density}
\title{Plot Imputation Density}
\usage{
plot_imputation_density(before, after)
}
\arguments{
\item{before}{A SummarizedExperiment object containing data before imputation.}

\item{after}{A SummarizedExperiment object containing data after imputation.}
}
\value{
A combined plot of intensity density before and after imputation.
}
\description{
This function creates density plots comparing the intensity distributions before and after imputation.
}
\examples{
# Example usage:
# p <- plot_imputation_density(before, after)
# print(p)
}
