% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_intensity_boxplot}
\alias{plot_intensity_boxplot}
\title{Plot Intensity Boxplot}
\usage{
plot_intensity_boxplot(MS1.df, MS2.df)
}
\arguments{
\item{MS1.df}{A data frame containing MS1 quantification data.}

\item{MS2.df}{A data frame containing MS2 quantification data.}
}
\value{
A combined plot of intensity boxplots for MS1 and MS2 data frames.
}
\description{
This function creates boxplots of log2 intensities for MS1 and MS2 data frames.
}
\examples{
# Example usage:
# p <- plot_intensity_boxplot(MS1.df, MS2.df)
# print(p)
}
