% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_missing}
\alias{plot_missing}
\title{Plot Missing Values}
\usage{
plot_missing(MS1.df, MS2.df)
}
\arguments{
\item{MS1.df}{A data frame containing MS1 quantification data.}

\item{MS2.df}{A data frame containing MS2 quantification data.}
}
\value{
A combined plot of missing values for MS1 and MS2 data frames.
}
\description{
This function creates bar plots showing the count of missing values in MS1 and MS2 data frames.
}
\examples{
# Example usage:
# p <- plot_missing(MS1.df, MS2.df)
# print(p)
}
