% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA}
\usage{
plot_pca(data, n = 500, maxPC = 2, center = TRUE, scale = TRUE, plot_all = F)
}
\arguments{
\item{data}{A SummarizedExperiment object containing quantification data.}

\item{n}{An integer specifying the number of top variable genes to use for PCA.}

\item{maxPC}{An integer specifying the maximum number of principal components to plot.}

\item{center}{A logical value indicating whether to center the data.}

\item{scale}{A logical value indicating whether to scale the data.}

\item{plot_all}{A logical value indicating whether to plot all principal components.}
}
\value{
PCA plots.
}
\description{
This function performs PCA on the data and creates PCA plots.
}
\examples{
# Example usage:
# plot_pca(data, n = 500, maxPC = 2)
}
