% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plot Volcano}
\usage{
plot_volcano(res, title = "", lfc_limit = NA)
}
\arguments{
\item{res}{A data frame containing differential expression results.}

\item{title}{A character string specifying the title of the plot.}

\item{lfc_limit}{A numeric value specifying the limit for log fold change.}
}
\value{
A volcano plot.
}
\description{
This function creates a volcano plot of differential expression results.
}
\examples{
# Example usage:
# plot_volcano(res, title = "Volcano Plot")
}
